# docs/swagger/projects.yaml
openapi: 3.0.0
info:
  title: Brian API - Projects Module
  version: 1.0.0
  description: API endpoints for managing projects, tasks, jobs, and related entities.
servers:
  - url: http://localhost:3000 # Replace with actual server URL if different
    description: Development server
...
paths:
  /api/v1/projects:
    get:
      tags:
        - Projects
      summary: List all projects
      description: Retrieves a list of projects with support for pagination, filtering, sorting, and field selection.
      parameters:
        - name: limit
          in: query
          description: Maximum number of projects to return per page.
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 10
        - name: offset
          in: query
          description: Number of projects to skip for pagination.
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: sort
          in: query
          description: Comma-separated list of fields to sort by (e.g., "name:asc,created_at:desc"). Allowed fields include id, name, hours, created_at, start_date, updated_at, closed_at, due_date, status_id, project_open, modified_at.
          required: false
          schema:
            type: string
            example: "created_at:desc"
        - name: fields
          in: query
          description: Comma-separated list of fields to include in the response (e.g., "id,name,status_id").
          required: false
          schema:
            type: string
            example: "id,name,start_date,due_date,status_id"
        # Filtering Parameters (add more as needed based on buildWhereClause)
        - name: name
          in: query
          description: Filter projects by name (case-insensitive, partial match).
          required: false
          schema:
            type: string
        - name: status_id
          in: query
          description: Filter projects by status ID (comma-separated for multiple values).
          required: false
          schema:
            type: string # Can be integer or comma-separated string
            example: "2,3"
        - name: project_open
          in: query
          description: Filter projects by open status (true/false).
          required: false
          schema:
            type: boolean
        - name: archived
          in: query
          description: Filter projects by archived status (true/false).
          required: false
          schema:
            type: boolean
        - name: company_id
          in: query
          description: Filter projects by company ID.
          required: false
          schema:
            type: integer
        - name: contract_id
          in: query
          description: Filter projects by contract ID.
          required: false
          schema:
            type: integer
        - name: created_by_user_id
          in: query
          description: Filter projects by creator user ID.
          required: false
          schema:
            type: integer
        - name: created_from
          in: query
          description: Filter projects created on or after this date (YYYY-MM-DD).
          required: false
          schema:
            type: string
            format: date
        - name: created_to
          in: query
          description: Filter projects created on or before this date (YYYY-MM-DD).
          required: false
          schema:
            type: string
            format: date
        # Add other date filters: start_from/to, due_from/to, closed_from/to, modified_from/to
      security:
        - BearerAuth: []
      responses:
        '200':
          description: A list of projects matching the criteria.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      total:
                        type: integer
                        description: Total number of projects matching the filter criteria.
                        example: 50
                      limit:
                        type: integer
                        description: The limit used for this request.
                        example: 10
                      offset:
                        type: integer
                        description: The offset used for this request.
                        example: 0
                      projects:
                        type: array
                        items:
                          $ref: '#/components/schemas/Project' # Reference to Project schema
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Projects
      summary: Create a new project
      description: Adds a new project to the system.
      security:
        - BearerAuth: []
      requestBody:
        description: Project data to create.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectInput'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/projects/{id}:
    get:
      tags:
        - Projects
      summary: Get a single project by ID
      description: Retrieves details for a specific project, with an option to include related data like tasks, jobs, and assigned people.
      parameters:
        - name: id
          in: path
          description: Numeric ID of the project to retrieve.
          required: true
          schema:
            type: integer
            format: int32
        - name: include
          in: query
          description: Comma-separated list of related data to include (e.g., "tasks,jobs,people,status,creator,modifier").
          required: false
          schema:
            type: string
            example: "tasks,people"
        - name: fields
          in: query
          description: Comma-separated list of project fields to include in the response.
          required: false
          schema:
            type: string
            example: "id,name,status_id"
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Project details retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Project' # Reference to Project schema, potentially with included data
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Projects
      summary: Update an existing project
      description: Modifies the details of a specific project by its ID.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Numeric ID of the project to update.
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: Project data to update. Only provided fields will be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectInput'
      responses:
        '200':
          $ref: '#/components/responses/Updated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Projects
      summary: Delete a project
      description: Removes a project from the system by its ID. May fail if dependencies exist (e.g., active tasks).
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Numeric ID of the project to delete.
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          $ref: '#/components/responses/Deleted'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/projects/{id}/tasks:
    get:
      tags:
        - Projects
      summary: List tasks for a specific project
      description: Retrieves a list of tasks associated with a given project ID, with pagination support.
      parameters:
        - name: id
          in: path
          description: Numeric ID of the project whose tasks are to be retrieved.
          required: true
          schema:
            type: integer
            format: int32
        - name: limit
          in: query
          description: Maximum number of tasks to return.
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: offset
          in: query
          description: Number of tasks to skip for pagination.
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        # Add filter/sort/fields parameters for tasks if implemented in service
      security:
        - BearerAuth: []
      responses:
        '200':
          description: A list of tasks for the specified project.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      total:
                        type: integer
                        description: Total number of tasks for this project.
                      limit:
                        type: integer
                      offset:
                        type: integer
                      tasks:
                        type: array
                        items:
                          $ref: '#/components/schemas/Task' # Reference to Task schema
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' # If project ID not found
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/projects/{id}/jobs:
    get:
      tags:
        - Projects
      summary: List jobs for a specific project
      description: Retrieves a list of jobs associated with a given project ID, with pagination support.
      parameters:
        - name: id
          in: path
          description: Numeric ID of the project whose jobs are to be retrieved.
          required: true
          schema:
            type: integer
            format: int32
        - name: limit
          in: query
          description: Maximum number of jobs to return.
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: offset
          in: query
          description: Number of jobs to skip for pagination.
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        # Add filter/sort/fields parameters for jobs if implemented
      security:
        - BearerAuth: []
      responses:
        '200':
          description: A list of jobs for the specified project.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      total:
                        type: integer
                        description: Total number of jobs for this project.
                      limit:
                        type: integer
                      offset:
                        type: integer
                      jobs:
                        type: array
                        items:
                          $ref: '#/components/schemas/Job' # Reference to Job schema
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' # If project ID not found
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/projects/{id}/people:
    get:
      tags:
        - Projects
      summary: List people assigned to a specific project
      description: Retrieves a list of people assigned to a given project ID, with pagination support.
      parameters:
        - name: id
          in: path
          description: Numeric ID of the project whose assigned people are to be retrieved.
          required: true
          schema:
            type: integer
            format: int32
        - name: limit
          in: query
          description: Maximum number of people to return.
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: offset
          in: query
          description: Number of people to skip for pagination.
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        # Add filter/sort/fields parameters for people if implemented
      security:
        - BearerAuth: []
      responses:
        '200':
          description: A list of people assigned to the specified project.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      total:
                        type: integer
                        description: Total number of people assigned to this project.
                      limit:
                        type: integer
                      offset:
                        type: integer
                      people:
                        type: array
                        items:
                          $ref: '#/components/schemas/Person' # Reference to Person schema
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' # If project ID not found
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
    Project:
      type: object
      properties:
        id:
          type: integer
          format: int32
          readOnly: true
        name:
          type: string
        hours:
          type: integer
          nullable: true
        contract_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        start_date:
          type: string
          format: date
          nullable: true
        company_id:
          type: integer
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true # Usually updated by system
        closed_at:
          type: string
          format: date-time
          nullable: true
        archived:
          type: boolean
          nullable: true
        due_date:
          type: string
          format: date
          nullable: true
        time_estimate_seconds:
          type: integer
          nullable: true
        project_category_id:
          type: integer
          nullable: true
        contract_value:
          type: integer # Assuming bigint maps to integer here, adjust if needed
          format: int64
          nullable: true
        material_cost:
          type: integer
          format: int64
          nullable: true
        customer_name_id:
          type: integer
          nullable: true
        product_category_id:
          type: integer
          nullable: true
        status_id:
          type: integer
          default: 2
        concrete_color_id:
          type: integer
          nullable: true
        address_id:
          type: integer
          nullable: true
        work_category:
          type: integer
          nullable: true
        project_open:
          type: boolean
          default: true
        clickup_task_id:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        created_by_user_id:
          type: integer
          nullable: true
          readOnly: true
        modified_by_user_id:
          type: integer
          nullable: true
          readOnly: true
        modified_at:
          type: string
          format: date-time
          readOnly: true
      example:
        id: 1
        name: "New Website Development"
        hours: 120
        contract_id: 5
        created_at: "2025-04-16T14:00:00Z"
        start_date: "2025-04-20"
        company_id: 2
        updated_at: "2025-04-16T15:30:00Z"
        closed_at: null
        archived: false
        due_date: "2025-07-31"
        time_estimate_seconds: 432000
        project_category_id: 1
        contract_value: 50000
        material_cost: 2000
        customer_name_id: 10
        product_category_id: 3
        status_id: 2
        concrete_color_id: null
        address_id: 15
        work_category: 1
        project_open: true
        clickup_task_id: "abc123xyz"
        notes: "Initial phase involves setting up the development environment."
        created_by_user_id: 1
        modified_by_user_id: 1
        modified_at: "2025-04-16T15:30:00Z"

    # Define Task, Job, Person schemas later
...
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            noToken:
              summary: No token provided
              value:
                success: false
                error: "Access denied. No token provided."
            invalidToken:
              summary: Invalid/Expired token
              value:
                success: false
                error: "Invalid token." # or "Token expired."
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token in the format 'Bearer <token>'
security:
  - BearerAuth: [] # Apply JWT authentication globally by default